# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "https-tunnel control plane",
    "version": "1.0.0",
    "summary": "Issues tunnel sessions and reports on them.",
    "description": "The control plane runs on the base domain, `tunnel.example.com`. It is the\nonly host that serves this API and the web UI.\n\nEvery other host under the base domain, `<label>.tunnel.example.com`, is a\ntunnel. Those hostnames serve whatever the connected client is running locally\nand have no API of their own, so nothing in this document applies to them. A\nrequest to a tunnel hostname whose client is offline is answered by the proxy\nitself with `502` and the offline page.\n\nA client uses two endpoints. `POST /api/v1/connect` registers or resumes a\nsession and returns the public URL, then `GET /api/v1/tunnel` upgrades the\nconnection to the tunnel protocol and keeps it open. The remaining endpoints\nare for inspection and cleanup.\n",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "https-tunnel",
      "url": "/docs"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server, whichever host is serving this document"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/connect": {
      "post": {
        "operationId": "connect",
        "tags": [
          "tunnel"
        ],
        "summary": "Register or resume a session",
        "description": "Returns the session id and the public URL the tunnel will be served on.\nPassing a `session_id` that the same API key owns resumes that session, keeping its URL; an unknown or foreign id is ignored and a fresh session is issued instead.\nA `subdomain_request` is granted when that label is free, and otherwise silently replaced by a random one, so always read the URL from the response rather than assuming the request was met.\nCalling this repeatedly without a `session_id` creates a new session each time, so clients should persist the id they get back.\n",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectRequest"
              },
              "examples": {
                "fresh": {
                  "summary": "Ask for a new session",
                  "value": {}
                },
                "resume": {
                  "summary": "Resume a known session",
                  "value": {
                    "session_id": "6wjulksypp3hbptpm5hqbn2davqbroyn"
                  }
                },
                "request_a_name": {
                  "summary": "Ask for a specific subdomain",
                  "value": {
                    "subdomain_request": "chris"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued or resumed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectResponse"
                },
                "example": {
                  "session": "6wjulksypp3hbptpm5hqbn2davqbroyn",
                  "url": "https://qr8xfcb946mw.tunnel.example.com"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "description": "No subdomain could be allocated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tunnel": {
      "get": {
        "operationId": "openTunnel",
        "tags": [
          "tunnel"
        ],
        "summary": "Upgrade to the tunnel protocol",
        "description": "Switches the connection from HTTP to the binary `https-tunnel` protocol and\nkeeps it open for the lifetime of the client. Attaching a second tunnel to\nthe same session displaces the first, so a reconnecting client always wins\nover a socket the server has not yet noticed is dead.\n\nAfter the `101` the connection is no longer HTTP, so the framing below is\noutside what OpenAPI can describe. Each frame is a one byte type, an eight\nbyte big endian stream id, a four byte big endian payload length, then the\npayload. Stream id 0 carries hello, ping and pong; every proxied request\ngets its own id, allocated by the server, and its head, body and end frames\nall share it. Frame types: 1 hello, 2 request head, 3 request body,\n4 request end, 5 response head, 6 response body, 7 response end, 8 abort,\n9 ping, 10 pong. Head frames carry JSON, body frames carry raw bytes split\nat 1 MiB.\n",
        "parameters": [
          {
            "name": "X-Tunnel-Session",
            "in": "header",
            "required": true,
            "description": "Session id returned by `POST /api/v1/connect`.",
            "schema": {
              "type": "string"
            },
            "example": "6wjulksypp3hbptpm5hqbn2davqbroyn"
          },
          {
            "name": "Connection",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "Upgrade"
            }
          },
          {
            "name": "Upgrade",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "https-tunnel"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Protocol switched. The server immediately sends a hello frame.",
            "headers": {
              "Upgrade": {
                "schema": {
                  "type": "string",
                  "const": "https-tunnel"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "const": "Upgrade"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Unknown session, or a session belonging to another key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "426": {
            "description": "The request did not ask for the `https-tunnel` upgrade.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <api key>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to the bearer header; the same key either way."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tunnel_admin",
        "description": "Session cookie set by /api/v1/auth/login. The web UI uses this; API clients use a key."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or unrecognized API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "invalid api key"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ConnectRequest": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string",
            "description": "Resume this session instead of issuing a new one."
          },
          "client_info": {
            "type": "string",
            "description": "Free text recorded in the server log."
          },
          "subdomain_request": {
            "type": "string",
            "description": "A label the client would like, so \"chris\" asks for https://chris.<base domain>/. It is granted when the label is free, or when another session of the same API key holds it. A label held by a different key, or a reserved one such as www or api, is refused quietly and a random label is issued instead, so the URL in the response is the authoritative answer. Sanitised to lowercase letters, digits and hyphens, and ignored entirely when the API key pins a fixed label.",
            "pattern": "^[a-z0-9][a-z0-9-]{0,39}$"
          }
        },
        "additionalProperties": false
      },
      "ConnectResponse": {
        "type": "object",
        "required": [
          "session",
          "url"
        ],
        "properties": {
          "session": {
            "type": "string",
            "description": "Session id. Persist it to reclaim the same URL later."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Public URL the tunnel is served on. It is built from the session's subdomain, the server's base domain and the scheme the request arrived over, so a server behind a TLS terminating proxy reports https even though it listens on plain HTTP."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}